refactor(mangareader): drop the title clone and keep helper private
